Create Authenticated Layout

+++ database/migrations/2025_03_04_221203_create_detail_layanans_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('detail_layanans', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->float('harga');
+            $table->integer('harga');
             $table->enum('status_pengajuan',['Menunggu Persetujuan','Disetujui']);
             $table->foreignId('fk_layanan')->constrained('layanans','id')->delete('cascade');
             $table->foreignId('fk_pet_house')->constrained('pet_houses','id')->delete('cascade');
